Clean up NewCredentials
