+++ packages/teleport/src/Welcome/NewCredentials/TwoFaInfo.tsx
